pyproject.toml
[project]
name = "arxiv-paper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
